# Replace exponential function in gray source calculations